// infrafi-web/src/hooks/buy-stake/types.ts
'use client'

export enum StakeMode {
  Stake = 'stake',
  Unstake = 'unstake',
}

export enum BuyMode {
  Buy = 'buy',
  Withdraw = 'withdraw',
}

export enum UnstakeFooterState {
  WindowInfo = 'window-info',
  ScheduleLink = 'schedule-link',
}

export enum StakeToken {
  USDtel = 'USDtel',
  SUSDtel = 'sUSDtel',
}